+++ CHANGES.rst
@@ -5,3 +5,4 @@ Development
 -----------
 
 * Initial setup
+* Naive implementation of the server-side Google authentication flow